+++ bloom_nofos/nofos/templatetags/get_icon_for_section.py
@@ -6,5 +6,5 @@ from .utils import get_icon_for_section as get_icon_for_section_func
 
 
 @register.filter()
-def get_icon_for_section(section_name=""):
-    return get_icon_for_section_func(section_name)
+def get_icon_for_section(section_name="", theme=""):
+    return get_icon_for_section_func(section_name, theme)