+++ CHANGELOG.md
@@ -11,6 +11,10 @@ The format is based on Keep a Changelog, and this project adheres to Semantic Ve
 - Added icons to the section title pages
 - Add favicon
 
+### Fixed
+
+- View page looks more like rendered NOFO
+
 ## [0.0.2] - 2023-12-30
 
 ### Added