+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on Keep a Changelog, and this project adheres to Semantic Versioning.
 
+## [0.0.4] - 2023-01-03
+
+### Added
+
+- Added icons to the table of contents page
+
 ## [0.0.3] - 2023-01-02
 
 ### Added